requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
python-dotenv
scikit-learn
